Add SolidGroup for multiple objects
Multiple Objects can't be rendered without some kind of collection and a
more generic means of performing hit detection. This introduces a
"Solid" trait for defining what it means to be hit by a ray. It also
adds an implementation for Sphere and a "SolidGroup" for collecting sets
of geometry to be rendered.